refactored DB and middlewares made sure to change property names within card display todo: testing and dummy data
