--- server/client-error.js
@@ -1,8 +1,0 @@
-class ClientError {
-  constructor(status, message) {
-    this.status = status;
-    this.message = message;
-  }
-}
-
-module.exports = ClientError;
