+++ server/middlewares/error-middleware.js
@@ -1,5 +1,5 @@
 const { JsonWebTokenError } = require('jsonwebtoken');
-const ClientError = require('./client-error');
+const ClientError = require('../lib/client-error');
 
 function errorMiddleware(err, req, res, next) {
   if (err instanceof ClientError) {
